+++ script/workspace/workspace.lua
@@ -212,7 +212,6 @@ local function loadFileFactory(root, progressData, isLibrary)
                                     action = 'set',
                                     value  = config.get 'Lua.workspace.maxPreload'
                                            + math.max(1000, config.get 'Lua.workspace.maxPreload'),
-                                    uri    = m.uri,
                                 }
                             }
                         end