+++ changelog.md
@@ -4,6 +4,7 @@
 * `NEW` `VSCode`: click the status bar icon to operate:
     * run workspace diagnostics
 * `FIX` `debug.getuservalue` and `debug.setuservalue` should not exist in `Lua 5.1`
+* `FIX` infer of `class[][]`
 
 ## 2.2.3
 `2021-7-9`