+++ changelog.md
@@ -1,5 +1,8 @@
 # changelog
 
+## 2.3.1
+* `FIX` [#606](https://github.com/sumneko/lua-language-server/issues/606)
+
 ## 2.3.0
 `2021-7-16`
 * `NEW` `VSCode`: click the status bar icon to operate: